third_party/blink/web_tests/platform/mac/virtual/plz-dedicated-worker/external/wpt/fetch/private-network-access/service-worker-background-fetch.https.window-expected.txt

This is a testharness.js-based test.
FAIL local to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
FAIL private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
FAIL private to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
FAIL public to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
FAIL public to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
FAIL public to public: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError"
PASS treat-as-public to local: success.
PASS treat-as-public to private: success.
PASS treat-as-public to public: success.
Harness: the test ran to completion.
